Remove handle module
